Add text style reducer
